Style: Cleanup folder structure

+++ examples/run_cantilever.py
@@ -1,4 +1,4 @@
-### Mass Spring System simulation falling under gravity.
+### Cantilever beam simulation falling under gravity.
 
 import numpy as np
 import warp as wp
@@ -6,6 +6,9 @@ import warp as wp
 import matplotlib.pyplot as plt
 import argparse
 import time, os, shutil
+
+import sys
+sys.path.append("./src")
 
 from _utils import voxel2hex, hex2tets
 from _renderer import PbrtRenderer, export_mp4
@@ -20,18 +23,11 @@ plt.rcParams.update({"font.family": 'serif'})#, "font.serif": ['Computer Modern'
 mm = 1/25.4
 
 
-class MassSpringSim:
-    """
-    Mass Spring System Simulation, where the mass is standardized as a 1m x 1m x 1m cube of 1kg. 
-    """
-    def __init__ (self, nx=11, density=1, youngs_modulus=9e3, poissons_ratio=0.45, dx_tol=1e-9, max_iter=1000, device="cuda"):
+class CantileverSim:
+    def __init__ (self, nx=(10, 3, 3), dx=(0.01, 0.01, 0.01), density=1070, youngs_modulus=250e3, poissons_ratio=0.45, dx_tol=1e-9, max_iter=100, device="cuda"):
         ### Set up tetrahedral mesh
-        dx = 1.0/nx
-        voxels = np.ones((nx, nx, 2*nx), dtype=bool)
-        voxels[:,:,nx:] = False
-        voxels[int(nx//2):int(nx//2)+1,int(nx//2):int(nx//2)+1,nx:] = True
-        vertices, hex_elements = voxel2hex(voxels, dx, dx, dx)
-        vertices[:,:2] -= vertices.mean(axis=0)[:2]
+        voxels = np.ones(nx, dtype=bool)
+        vertices, hex_elements = voxel2hex(voxels, *dx)
         elements = hex2tets(hex_elements)
         n_vertices = vertices.shape[0]
         n_elements = elements.shape[0]
@@ -43,9 +39,9 @@ class MassSpringSim:
         active_mask = np.ones((vertices.shape[0],), dtype=bool)
         tip_idx = []
         for i in range(vertices.shape[0]):
-            if vertices[i,2] > 2*nx*dx-1e-3:
+            if vertices[i,0] < 1e-3:
                 active_mask[i] = False
-            elif vertices[i,2] < 1e-3:
+            elif vertices[i,0] > (nx[0] * dx[0] - 1e-3):
                 tip_idx.append(i)
         self.tip_idx = tip_idx
         print(f"Active vertices: {np.sum(active_mask)}/{vertices.shape[0]}")
@@ -71,11 +67,11 @@ class MassSpringSim:
             device=device
         )
 
-        ### Add a wall plane as a hex mesh box. Center should align with mass spring top center.
-        com = vertices.mean(axis=0)
-        wall_width, wall_depth, wall_height = 1.0, 1.0, 0.1
-        wall_translation = np.array([com[0]-wall_width/2, com[1]-wall_depth/2, 2*nx*dx-wall_height/2])
-        self.wall_vertices, self.wall_elements = voxel2hex(np.ones((1,1,1), dtype=bool), wall_width, wall_depth, wall_height)
+        ### Add a wall plane as a hex mesh box. Center should align with cantilever base center.
+        beam_com = vertices.mean(axis=0)
+        wall_width, wall_height, wall_depth = 0.02, 0.1, 0.1
+        wall_translation = np.array([-wall_width, beam_com[1]-wall_height/2, beam_com[2]-wall_depth/2])
+        self.wall_vertices, self.wall_elements = voxel2hex(np.ones((2,1,1), dtype=bool), wall_width/2, wall_height, wall_depth)
         self.wall_vertices += wall_translation
 
 
@@ -97,11 +93,11 @@ class MassSpringSim:
             'light_map': 'uffizi-large.exr',
             'sample': spp,
             'max_depth': 2,
-            'camera_pos': (0.0, -0.75, 0.3),   # Position of camera
+            'camera_pos': (0.3, -0.75, 0.3),   # Position of camera
             'camera_lookat': (0.0, 0.25, 0.1),     # Position that camera looks at
         }
         transforms=[
-            ('s', 0.1),
+            ('s', 2),
             ('t', [0, 0, 0.15])
         ]
         renderer = PbrtRenderer(options)
@@ -128,20 +124,26 @@ class MassSpringSim:
 
 def main (args):
     # Create output directories
-    output_folder = "outputs/mass_spring"
+    output_folder = "outputs/cantilever"
     os.makedirs(output_folder, exist_ok=True)
-    if os.path.isdir(f"{output_folder}/frames"):
-        shutil.rmtree(f"{output_folder}/frames")
-    os.makedirs(f"{output_folder}/frames", exist_ok=False)
+    if args.render:
+        if os.path.isdir(f"{output_folder}/frames"):
+            shutil.rmtree(f"{output_folder}/frames")
+        os.makedirs(f"{output_folder}/frames", exist_ok=False)
 
     # Set up simulation
-    sim = MassSpringSim(nx=args.nx, dx_tol=1e-5, device="cuda")
+    sim = CantileverSim(
+        nx=(args.nx, args.ny, args.nz), 
+        dx=(args.dx, args.dy, args.dz), 
+        density=1070, youngs_modulus=150e3, poissons_ratio=0.45,
+        dx_tol=1e-9, device="cuda"
+    )
 
     ### Begin the solve
-    n_seconds = 5.0
-    fps = 50
+    n_seconds = 1.5
+    fps = 100
     n_timesteps = int(n_seconds * fps)
-    n_substeps = 5
+    n_substeps = 10
     dt = 1/fps/n_substeps
 
     tip_positions = []
@@ -156,7 +158,7 @@ def main (args):
         print(f"---Timestep [{t:04d}/{n_timesteps}] ({1e3*dt*n_substeps:.1f}ms) in {1e3*(end_time - start_time):.3f}ms: Mean Positions: {solution.numpy().mean(axis=0)}")
 
         if args.render:
-            sim.render(solution.numpy(), filename=f"{output_folder}/frames/mass_spring_{t:03d}.png", spp=4)
+            sim.render(solution.numpy(), filename=f"{output_folder}/frames/cantilever_{t:03d}.png", spp=4)
 
         # Plot Tip Displacements
         fig, ax = plt.subplots(figsize=(3,2))
@@ -167,22 +169,31 @@ def main (args):
         ax.grid()
         ax.set_xlim(0, n_seconds)
         ax.ticklabel_format(axis='both', style='sci', scilimits=(0,0))
-        fig.savefig(f"{output_folder}/displacement_mass_spring.png", dpi=300, bbox_inches='tight')
+        fig.savefig(f"{output_folder}/displacement_cantilever.png", dpi=300, bbox_inches='tight')
         plt.close(fig)
 
     # Store as csv
-    np.savetxt(f"{output_folder}/displacement_mass_spring.csv", np.array(tip_positions), delimiter=",")
+    np.savetxt(f"{output_folder}/displacement_cantilever.csv", np.array(tip_positions), delimiter=",")
 
     if args.render:
         # Render mp4
-        export_mp4(f"{output_folder}/frames", f"{output_folder}/mass_spring.mp4", fps=fps)
+        export_mp4(f"{output_folder}/frames", f"{output_folder}/cantilever.mp4", fps=int(0.25*fps))
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument("--nx", type=int, default=11, help="Number of voxels in x direction")
+    parser.add_argument("--nx", type=int, default=10, help="Number of voxels in x direction")
+    parser.add_argument("--ny", type=int, default=3, help="Number of voxels in y direction")
+    parser.add_argument("--nz", type=int, default=3, help="Number of voxels in z direction")
     parser.add_argument("--dx", type=float, default=0.01, help="Voxel size in x direction")
+    parser.add_argument("--dy", type=float, default=None, help="Voxel size in y direction")
+    parser.add_argument("--dz", type=float, default=None, help="Voxel size in z direction")
     parser.add_argument("--render", action='store_true', help="Visualize option")
     args = parser.parse_args()
+
+    if args.dy is None:
+        args.dy = args.dx
+    if args.dz is None:
+        args.dz = args.dx
     
     main(args)
